--- Assets/KGW_Work/ClearSceneUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6d7b2cde018bf10428ffce51a07d88a2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
