+++ Assets/KGW_Work/Frefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 54436e5867403f64a93b30effd59f1d5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
