--- Assets/KGW_Work/MapIntroMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e2d881df46286594f956f5f4746818e3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
