+++ Assets/KGW_Work/Movie_004.mp4.meta
@@ -1,0 +1,18 @@
+fileFormatVersion: 2
+guid: 7f2095c12ed4b7b42b0e4d2d6728a824
+VideoClipImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  frameRange: 0
+  startFrame: -1
+  endFrame: -1
+  colorSpace: 0
+  deinterlace: 0
+  encodeAlpha: 0
+  flipVertical: 0
+  flipHorizontal: 0
+  importAudio: 1
+  targetSettings: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
