--- Assets/KGW_Work/RankView.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6789d236cfce3fd419ae9b4e5b2e4bc0
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
