+++ Assets/KGW_Work/Scripts/SaveManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 66b3215062c2fee42b884bc9029059ad
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
